Structural Change - remove_method --> remove_possible_method, extract #redefine_method in other file
