+++ actionpack/lib/action_dispatch/routing/mapper.rb
@@ -3,7 +3,7 @@ require 'active_support/core_ext/hash/reverse_merge'
 require 'active_support/core_ext/hash/slice'
 require 'active_support/core_ext/enumerable'
 require 'active_support/core_ext/array/extract_options'
-require 'active_support/core_ext/module/remove_method'
+require 'active_support/core_ext/module/redefine_method'
 require 'active_support/inflector'
 require 'action_dispatch/routing/redirection'
