+++ actionpack/lib/action_dispatch/routing/route_set.rb
@@ -3,7 +3,8 @@ require 'forwardable'
 require 'thread_safe'
 require 'active_support/core_ext/object/to_query'
 require 'active_support/core_ext/hash/slice'
-require 'active_support/core_ext/module/remove_method'
+require 'active_support/core_ext/module/redefine_method'
+require 'active_support/core_ext/module/remove_possible_method'
 require 'active_support/core_ext/array/extract_options'
 require 'action_controller/metal/exceptions'
