+++ actionview/lib/action_view/layouts.rb
@@ -1,5 +1,5 @@
 require "action_view/rendering"
-require "active_support/core_ext/module/remove_method"
+require "active_support/core_ext/module/remove_possible_method"
 
 module ActionView
   # Layouts reverse the common pattern of including shared headers and footers in many templates to isolate changes in
