+++ actionview/lib/action_view/lookup_context.rb
@@ -1,5 +1,5 @@
 require 'thread_safe'
-require 'active_support/core_ext/module/remove_method'
+require 'active_support/core_ext/module/remove_possible_method'
 require 'active_support/core_ext/module/attribute_accessors'
 require 'action_view/template/resolver'
