+++ actionview/lib/action_view/test_case.rb
@@ -1,4 +1,4 @@
-require 'active_support/core_ext/module/remove_method'
+require 'active_support/core_ext/module/remove_possible_method'
 require 'action_controller'
 require 'action_controller/test_case'
 require 'action_view'
