+++ activerecord/lib/active_record/associations.rb
@@ -1,6 +1,6 @@
 require 'active_support/core_ext/enumerable'
 require 'active_support/core_ext/string/conversions'
-require 'active_support/core_ext/module/remove_method'
+require 'active_support/core_ext/module/remove_possible_method'
 require 'active_record/errors'
 
 module ActiveRecord
