+++ activesupport/lib/active_support/core_ext/class/attribute.rb
@@ -1,5 +1,5 @@
 require 'active_support/core_ext/kernel/singleton_class'
-require 'active_support/core_ext/module/remove_method'
+require 'active_support/core_ext/module/remove_possible_method'
 require 'active_support/core_ext/array/extract_options'
 
 class Class
