+++ activesupport/lib/active_support/core_ext/date/conversions.rb
@@ -1,7 +1,7 @@
 require 'date'
 require 'active_support/inflector/methods'
 require 'active_support/core_ext/date/zones'
-require 'active_support/core_ext/module/remove_method'
+require 'active_support/core_ext/module/remove_possible_method'
 
 class Date
   DATE_FORMATS = {
