+++ activesupport/lib/active_support/core_ext/module.rb
@@ -7,5 +7,6 @@ require 'active_support/core_ext/module/attr_internal'
 require 'active_support/core_ext/module/concerning'
 require 'active_support/core_ext/module/delegation'
 require 'active_support/core_ext/module/deprecation'
-require 'active_support/core_ext/module/remove_method'
+require 'active_support/core_ext/module/remove_possible_method'
 require 'active_support/core_ext/module/qualified_const'
+require 'active_support/core_ext/module/redefine_method'
