+++ activesupport/lib/active_support/core_ext/module/remove_possible_method.rb
@@ -4,9 +4,4 @@ class Module
       undef_method(method)
     end
   end
-
-  def redefine_method(method, &block)
-    remove_possible_method(method)
-    define_method(method, &block)
-  end
 end
